feat(lnurlpay): pass lnurlDetails to navState #1216
